+++ Nos/Extensions/Localization+Nos.swift
@@ -31,8 +31,9 @@ extension Localizable {
             return ""
         }
     }
-
-    func localizedText(_ arguments: [String: String]) -> LocalizedStringKey {
+  
+    /// Use this function in place of `text(_ arguments:[String: String]) -> String` to use Markdown-formatted text
+    func localizedMarkdown(_ arguments: [String: String]) -> LocalizedStringKey {
         LocalizedStringKey(text(arguments))
     }