+++ Nos/Extensions/Localization+Nos.swift
@@ -31,7 +31,7 @@ extension Localizable {
             return ""
         }
     }
-
+  
     /// Use this function in place of `text(_ arguments:[String: String]) -> String` to use Markdown-formatted text
     func localizedMarkdown(_ arguments: [String: String]) -> LocalizedStringKey {
         LocalizedStringKey(text(arguments))