+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add list of followers and relays in the Profile screen.
+- Add a loading placeholder for note contents.
 - Added automatic file uploads to nostr.build.
 
 ## [0.1 (60)] - 2023-08-01Z